+++ backend-api/src/main/java/com/project/backend_api/services/BookService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface BookService {
     public String createBook(Book book);
     public String updateBook(Book book);
-    public String deleteBook(int bookId);
-    public Book getBook(int bookId);
+    public String deleteBook(Long bookId);
+    public Book getBook(Long bookId);
     public List<BookDTO> getAllBooks();
 }
